style and fix user profile states
