+++ frontend/src/components/Header.js
@@ -165,7 +165,7 @@ export default function Header(props) {
             shadowOffset: { width: 0, height: 2 },
             shadowOpacity: 0.8,
             shadowRadius: 2,  
-            elevation: 5
+            elevation: onMainNav ? 5 : 0
 
         }, 
         title: {
