+++ frontend/src/pages/off-nav/UserProfilePage.js
@@ -12,7 +12,7 @@ export default function UserProfilePage({ navigation, route }) {
 
     const {windowWidth, windowHeight} = useContext(DimensionContext)
 
-    const topMarginSize = windowWidth*0.1;
+    const topMarginSize = windowWidth * 0.1;
     const settingMarginSides = windowWidth * 0.05;
     const settingMarginTopBottom = windowWidth * 0.03;
     const descMargin = windowWidth * 0.03;
@@ -41,7 +41,7 @@ export default function UserProfilePage({ navigation, route }) {
         marginLeft: descMargin
       },
       settingsContainer: {
-        width: '100%',
+        width: windowWidth,
         marginTop: settingMarginTopBottom,
         borderTopWidth: 1,
         borderColor: '#D4D4D4',
@@ -49,14 +49,17 @@ export default function UserProfilePage({ navigation, route }) {
     })
 
     return (
-        <View style={{display: 'flex', alignItems: 'center', marginTop: topMarginSize}}>
+        <View style={{display: 'flex', alignItems: 'center'}}>
           <Header routeName={routeName} navigation={navigation} onMainNav={false}/>
-          <UserProfile
-            routeName={'User Profile Page'}
-            profilePicture={pfp}
-            userFirstLast={'Dummy One'}
-            username={'iamnumberone'}
-          />
+          
+          <View style={{marginTop: topMarginSize}}>
+              <UserProfile
+              routeName={'User Profile Page'}
+              profilePicture={pfp}
+              userFirstLast={'Dummy One'}
+              username={'iamnumberone'}
+            />
+            
 
           {/* Settings section */}
           <View style={styles.settingsContainer}>
@@ -66,7 +69,7 @@ export default function UserProfilePage({ navigation, route }) {
                 <Image 
                   source={require('assets/resetPassword.png')}
                   style={styles.icon}
-                />
+                  />
                 <Text style={styles.desc}>Reset Password</Text>
               </View>
               <Text>{">"}</Text>
@@ -75,14 +78,15 @@ export default function UserProfilePage({ navigation, route }) {
             <TouchableOpacity style={styles.setting}>
               <View style={styles.iconDescGroup}>
                 <Image 
-                  source={require('assets/resetPassword.png')}
+                  source={require('assets/deleteAccount.png')}
                   style={styles.icon}
-                />
+                  />
                 <Text style={styles.desc}>Delete Account</Text>
               </View>
               <Text>{">"}</Text>
             </TouchableOpacity>
           </View>
         </View>
+      </View>
     )
 }
